Move suggested folder model classes to separate file.
